Refactor CLI storage and plan selection
