+++ devcrew/cli.py
@@ -7,19 +7,25 @@ import json
 import logging
 import os
 import sys
-from dataclasses import dataclass
 from datetime import datetime
 from pathlib import Path
 from typing import Any, Optional
-from uuid import uuid4
 
 from dotenv import load_dotenv
 from langchain.chat_models import init_chat_model
 from langchain_core.language_models.chat_models import BaseChatModel
 from langchain_openai import ChatOpenAI
 
-import yaml
 from .llm import SafeLLM
+from .plan_selection import ensure_prompt_for_plan, select_saved_plan
+from .plan_storage import (
+    dump_plan_on_failure,
+    initialise_plan_storage,
+    persist_plan_result,
+    plan_to_yaml_text,
+    prepare_output_dir,
+    update_plan_metadata,
+)
 
 try:  # Optional streaming dependencies (not present in all crewAI installs)
     from crewai.events.base_event_listener import BaseEventListener
@@ -150,474 +156,30 @@ def build_orchestrator(args: argparse.Namespace) -> DynamicCrewOrchestrator:
     )
 
 
-def _prepare_output_dir(path: str) -> Path:
-    outdir = Path(path)
-    outdir.mkdir(parents=True, exist_ok=True)
-    return outdir
+def _prompt_for_execution(plan: CrewPlan) -> bool:
+    """Ask the user whether the freshly built crew should run."""
 
+    if not sys.stdin.isatty():
+        LOGGER.info("Non-interactive terminal detected; skipping execution.")
+        return False
 
-def _yaml_dump(path: Path, payload: Any) -> None:
-    with path.open("w", encoding="utf-8") as fh:
-        yaml.safe_dump(payload, fh, sort_keys=False, allow_unicode=True)
-
-
-def _yaml_load(path: Path) -> Any:
-    with path.open("r", encoding="utf-8") as fh:
-        return yaml.safe_load(fh)
-
-
-def _plan_to_yaml_payload(plan: CrewPlan) -> dict[str, Any]:
-    agents = [agent.model_dump() for agent in plan.agents]
-    tasks = [task.model_dump() for task in plan.tasks]
-    return {
-        "crew": {
-            "summary": plan.summary,
-            "process": plan.process,
-            "agents": [a.get("name") for a in agents],
-            "tasks": [t.get("name") for t in tasks],
-        },
-        "agents": agents,
-        "tasks": tasks,
-    }
-
-
-def _plan_to_yaml_text(plan: CrewPlan) -> str:
-    return yaml.safe_dump(
-        _plan_to_yaml_payload(plan), sort_keys=False, allow_unicode=True
-    )
-
-
-def _update_plan_metadata(plan_dir: Path, **updates: Any) -> None:
-    metadata_path = plan_dir / "metadata.yaml"
-    try:
-        existing = (
-            _yaml_load(metadata_path)
-            if metadata_path.exists()
-            else {}
-        )
-        metadata = existing if isinstance(existing, dict) else {}
-    except Exception:  # pragma: no cover - defensive IO handling
-        metadata = {}
-
-    clean_updates = {k: v for k, v in updates.items() if v is not None}
-    if not clean_updates and metadata_path.exists():
-        # Ensure the file exists even when there is nothing to update.
-        return
-
-    metadata.update(clean_updates)
-    _yaml_dump(metadata_path, metadata)
-
-
-def _write_plan_bundle(plan_dir: Path, plan: CrewPlan, prompt: Optional[str]) -> None:
-    bundle = _plan_to_yaml_payload(plan)
-    plan_dir.mkdir(parents=True, exist_ok=True)
-
-    _yaml_dump(plan_dir / "plan.yaml", bundle)
-    _yaml_dump(plan_dir / "agents.yaml", {"agents": bundle["agents"]})
-    _yaml_dump(plan_dir / "tasks.yaml", {"tasks": bundle["tasks"]})
-    _yaml_dump(plan_dir / "crew.yaml", bundle["crew"])
-
-    _update_plan_metadata(
-        plan_dir,
-        saved_at=datetime.now().isoformat(timespec="seconds"),
-        prompt=prompt,
-    )
-
-
-def _extract_section(data: Any, key: str) -> list[dict[str, Any]]:
-    if isinstance(data, dict):
-        section = data.get(key, [])
-    else:
-        section = data
-
-    if section is None:
-        return []
-    if isinstance(section, list):
-        return section
-
-    raise ValueError(f"Unsupported {key} structure in saved plan: {type(section)!r}")
-
-
-def _read_plan_bundle(plan_dir: Path) -> tuple[CrewPlan, Optional[str]]:
-    try:
-        plan_data = _yaml_load(plan_dir / "plan.yaml")
-    except FileNotFoundError:
-        plan_data = None
-
-    if isinstance(plan_data, dict):
-        agents_data = _extract_section(plan_data.get("agents"), "agents")
-        tasks_data = _extract_section(plan_data.get("tasks"), "tasks")
-        crew_data = plan_data.get("crew", {})
-    else:
-        agents_data = _extract_section(_yaml_load(plan_dir / "agents.yaml"), "agents")
-        tasks_data = _extract_section(_yaml_load(plan_dir / "tasks.yaml"), "tasks")
-        crew_data = _yaml_load(plan_dir / "crew.yaml") or {}
-
-    payload = {
-        "summary": crew_data.get("summary", ""),
-        "process": crew_data.get("process", "sequential"),
-        "agents": agents_data,
-        "tasks": tasks_data,
-    }
-
-    plan = CrewPlan.model_validate(payload)
-
-    prompt: Optional[str] = None
-    metadata_path = plan_dir / "metadata.yaml"
-    if metadata_path.exists():
-        try:
-            metadata = _yaml_load(metadata_path) or {}
-            if isinstance(metadata, dict):
-                raw_prompt = metadata.get("prompt")
-                if isinstance(raw_prompt, str) and raw_prompt.strip():
-                    prompt = raw_prompt
-        except Exception as exc:  # pragma: no cover - defensive
-            logging.warning("Failed to parse metadata for %s: %s", plan_dir, exc)
-
-    return plan, prompt
-
-
-@dataclass
-class SavedPlanRecord:
-    path: Path
-    plan: CrewPlan
-    prompt: Optional[str]
-    created_at: datetime
-    storage_format: str = "json"
-
-
-def _format_plan_label(record: SavedPlanRecord) -> str:
-    timestamp = record.created_at.strftime("%Y-%m-%d %H:%M:%S")
-    summary = record.plan.summary or "(no summary)"
-    return f"{timestamp} · {record.path.name} · {summary}"
-
-
-def _discover_saved_plans(outdir: Path) -> list[SavedPlanRecord]:
-    plans: list[SavedPlanRecord] = []
-
-    for plan_dir in sorted(
-        (p for p in outdir.glob("plan_*") if p.is_dir()),
-        key=lambda p: p.stat().st_mtime,
-        reverse=True,
-    ):
-        try:
-            plan, prompt = _read_plan_bundle(plan_dir)
-        except Exception as exc:  # pragma: no cover - defensive IO handling
-            logging.warning("Could not read saved plan bundle %s: %s", plan_dir, exc)
-            continue
-
-        created_at = datetime.fromtimestamp(plan_dir.stat().st_mtime)
-        plans.append(
-            SavedPlanRecord(
-                path=plan_dir,
-                plan=plan,
-                prompt=prompt,
-                created_at=created_at,
-                storage_format="yaml",
-            )
-        )
-
-    for plan_path in sorted(
-        outdir.glob("plan_*.json"), key=lambda p: p.stat().st_mtime, reverse=True
-    ):
-        try:
-            with plan_path.open("r", encoding="utf-8") as fh:
-                raw_data = json.load(fh)
-        except Exception as exc:  # pragma: no cover - defensive IO handling
-            logging.warning("Could not read saved plan %s: %s", plan_path, exc)
-            continue
-
-        plan_data: Optional[dict[str, Any]] = None
-        prompt: Optional[str] = None
-
-        if isinstance(raw_data, dict):
-            if "plan" in raw_data and isinstance(raw_data["plan"], dict):
-                plan_data = raw_data["plan"]
-                prompt = raw_data.get("prompt") or raw_data.get("_prompt")
-            else:
-                prompt = raw_data.get("_prompt")
-                if "_prompt" in raw_data:
-                    plan_data = {k: v for k, v in raw_data.items() if k != "_prompt"}
-                else:
-                    plan_data = raw_data
-
-        if not plan_data:
-            logging.warning("Ignoring malformed plan file: %s", plan_path)
-            continue
-
-        try:
-            plan = CrewPlan.model_validate(plan_data)
-        except Exception as exc:  # pragma: no cover - validation is already tested elsewhere
-            logging.warning("Failed to parse plan %s: %s", plan_path, exc)
-            continue
-
-        created_at = datetime.fromtimestamp(plan_path.stat().st_mtime)
-        plans.append(
-            SavedPlanRecord(
-                path=plan_path,
-                plan=plan,
-                prompt=prompt,
-                created_at=created_at,
-                storage_format="json",
-            )
-        )
-
-    return plans
-
-
-def _select_plan_with_curses(
-    plans: list[SavedPlanRecord],
-) -> tuple[Optional[SavedPlanRecord], bool]:
-    try:
-        import curses
-    except Exception:  # pragma: no cover - curses missing on some platforms
-        return None, False
-
-    if not plans:
-        return None, False
-
-    selected: dict[str, int] = {"index": 0}
-
-    def _menu(stdscr):
-        curses.curs_set(0)
-        idx = selected["index"]
-        offset = 0
-
-        while True:
-            stdscr.erase()
-            height, width = stdscr.getmaxyx()
-            instructions = "Use ↑/↓ to select a plan. Enter executes, q cancels."
-            stdscr.addnstr(0, 0, instructions, max(0, width - 1))
-
-            visible_rows = max(1, height - 2)
-            if idx < offset:
-                offset = idx
-            elif idx >= offset + visible_rows:
-                offset = idx - visible_rows + 1
-
-            for row in range(visible_rows):
-                plan_idx = offset + row
-                if plan_idx >= len(plans):
-                    break
-                record = plans[plan_idx]
-                label = _format_plan_label(record)
-                prefix = "➤ " if plan_idx == idx else "  "
-                text = (prefix + label)[: max(0, width - 1)]
-                if plan_idx == idx:
-                    stdscr.attron(curses.A_REVERSE)
-                    stdscr.addnstr(row + 1, 0, text, max(0, width - 1))
-                    stdscr.attroff(curses.A_REVERSE)
-                else:
-                    stdscr.addnstr(row + 1, 0, text, max(0, width - 1))
-
-            key = stdscr.getch()
-            if key in (curses.KEY_UP, ord("k")):
-                idx = (idx - 1) % len(plans)
-            elif key in (curses.KEY_DOWN, ord("j")):
-                idx = (idx + 1) % len(plans)
-            elif key in (curses.KEY_ENTER, 10, 13):
-                selected["index"] = idx
-                return
-            elif key in (27, ord("q")):
-                raise KeyboardInterrupt
-
-    try:
-        import curses
-
-        curses.wrapper(_menu)
-    except KeyboardInterrupt:
-        return None, True
-    except curses.error:  # pragma: no cover - terminal limitations
-        return None, False
-
-    return plans[selected.get("index", 0)], False
-
-
-def _select_plan_via_input(
-    plans: list[SavedPlanRecord],
-) -> tuple[Optional[SavedPlanRecord], bool]:
-    if not plans:
-        return None, False
-
-    print("\nSaved plans available:")
-    for idx, record in enumerate(plans, start=1):
-        label = _format_plan_label(record)
-        print(f" {idx}. {label}")
+    summary = plan.summary or "the planned crew"
+    print(f"\nPlan ready for: {summary}")
 
     while True:
         try:
-            response = input("Select a plan number to execute (empty to cancel): ")
+            response = input("Execute the crew now? [y/N]: ")
         except EOFError:
-            return None, True
+            LOGGER.info("No confirmation received; skipping execution.")
+            return False
 
-        if not response.strip():
-            return None, True
+        answer = response.strip().lower()
+        if answer in {"", "n", "no"}:
+            return False
+        if answer in {"y", "yes"}:
+            return True
 
-        if response.strip().lower() in {"q", "quit", "exit"}:
-            return None, True
-
-        try:
-            index = int(response.strip()) - 1
-        except ValueError:
-            print("Invalid selection. Please enter a number from the list.")
-            continue
-
-        if 0 <= index < len(plans):
-            return plans[index], False
-
-        print("Selection out of range. Try again.")
-
-
-def _select_saved_plan(outdir: Path) -> tuple[Optional[SavedPlanRecord], bool]:
-    plans = _discover_saved_plans(outdir)
-    if not plans:
-        logging.error("No saved plans were found in %s", outdir)
-        return None, False
-
-    if sys.stdin.isatty() and sys.stdout.isatty():
-        selection, cancelled = _select_plan_with_curses(plans)
-        if selection:
-            return selection, False
-        if cancelled:
-            logging.info("Plan selection cancelled by user.")
-            return None, True
-
-    if not sys.stdin.isatty():
-        logging.error("Interactive selection requires a TTY. Cannot continue.")
-        return None, False
-
-    selection, cancelled = _select_plan_via_input(plans)
-    if selection:
-        return selection, False
-    if cancelled:
-        logging.info("Plan selection cancelled by user.")
-        return None, True
-
-    return None, False
-
-
-def _ensure_prompt_for_plan(record: SavedPlanRecord) -> Optional[str]:
-    if record.prompt:
-        return record.prompt
-
-    if record.storage_format == "yaml" and record.path.is_dir():
-        metadata_path = record.path / "metadata.yaml"
-        if metadata_path.exists():
-            try:
-                metadata = _yaml_load(metadata_path) or {}
-                if isinstance(metadata, dict):
-                    stored_prompt = metadata.get("prompt")
-                    if isinstance(stored_prompt, str) and stored_prompt.strip():
-                        record.prompt = stored_prompt
-                        return stored_prompt
-            except Exception as exc:  # pragma: no cover - defensive IO handling
-                logging.warning("Failed to read prompt metadata for %s: %s", record.path, exc)
-
-    if not sys.stdin.isatty():
-        logging.error(
-            "The selected plan (%s) does not include the original prompt and no TTY is available to ask for it.",
-            record.path,
-        )
-        return None
-
-    try:
-        response = input(
-            f"Enter the original prompt for {record.path.name} (required to execute the crew): "
-        )
-    except EOFError:
-        logging.error("Prompt capture interrupted. Aborting execution.")
-        return None
-
-    prompt = response.strip()
-    if not prompt:
-        logging.error("A prompt is required to run the selected crew.")
-        return None
-
-    _persist_prompt_for_plan(record, prompt)
-    record.prompt = prompt
-
-    return prompt
-
-
-def _persist_prompt_for_plan(record: SavedPlanRecord, prompt: str) -> None:
-    timestamp = datetime.now().isoformat(timespec="seconds")
-
-    if record.storage_format == "yaml" and record.path.is_dir():
-        _update_plan_metadata(
-            record.path,
-            prompt=prompt,
-            updated_at=timestamp,
-        )
-        return
-
-    if record.path.is_file() and record.path.suffix == ".json":
-        try:
-            with record.path.open("r", encoding="utf-8") as fh:
-                raw = json.load(fh)
-            if isinstance(raw, dict):
-                raw["_prompt"] = prompt
-                with record.path.open("w", encoding="utf-8") as fh:
-                    json.dump(raw, fh, indent=2, ensure_ascii=False)
-        except Exception as exc:  # pragma: no cover - defensive IO handling
-            logging.warning("Failed to persist prompt for %s: %s", record.path, exc)
-
-
-def _initialise_plan_storage(
-    outdir: Path,
-    plan: CrewPlan,
-    prompt: Optional[str],
-    *,
-    source: Optional[Path] = None,
-) -> Path:
-    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-    suffix = uuid4().hex[:6]
-    plan_dir = outdir / f"plan_{timestamp}_{suffix}"
-
-    _write_plan_bundle(plan_dir, plan, prompt)
-
-    now_iso = datetime.now().isoformat(timespec="seconds")
-    _update_plan_metadata(
-        plan_dir,
-        status="planned",
-        executed=False,
-        plan_id=plan_dir.name,
-        source=str(source) if source else None,
-        initialised_at=now_iso,
-        updated_at=now_iso,
-    )
-    logging.info("Plan saved in %s", plan_dir)
-    return plan_dir
-
-
-def _persist_plan_result(
-    plan_dir: Path,
-    plan: CrewPlan,
-    result: Any,
-    fmt: str,
-    status: str,
-) -> Path:
-    plan_dir.mkdir(parents=True, exist_ok=True)
-
-    if fmt == "json":
-        result_file = plan_dir / "result.json"
-        out = {
-            "summary": plan.summary,
-            "process": plan.process,
-            "agents": [a.name for a in plan.agents],
-            "tasks": [t.name for t in plan.tasks],
-            "result": result,
-        }
-        with result_file.open("w", encoding="utf-8") as fh:
-            json.dump(out, fh, indent=2, ensure_ascii=False)
-    else:
-        result_file = plan_dir / "result.txt"
-        with result_file.open("w", encoding="utf-8") as fh:
-            if isinstance(result, (dict, list)):
-                fh.write(json.dumps(result, indent=2, ensure_ascii=False))
-            else:
-                fh.write(str(result))
-
-    logging.info("Outputs saved: %s , %s", plan_file, result_file)
+        print("Please answer 'y' or 'n'.")
 
 
 def _register_stream_listener(args: argparse.Namespace) -> None:
@@ -710,25 +272,10 @@ async def main(argv: Optional[list[str]] = None) -> int:
     result_persisted = False
 
     try:
-        # Registrar listener de streaming si procede
-        if args.stream and BaseEventListener is not object:
-            if TokenStreamListener is None:
-                logging.warning(
-                    "Streaming requested but TokenStreamListener is unavailable; skipping listener registration."
-                )
-            else:
-                listener = TokenStreamListener()
-                if callable(_get_bus):
-                    bus = _get_bus()
-                    # Algunas versiones requieren .add_listener, otras instanciar el listener basta
-                    try:
-                        bus.add_listener(listener)  # type: ignore[attr-defined]
-                    except Exception:
-                        # Si el bus inyecta listeners automáticamente vía entrypoints, ignoramos
-                        pass
+        _register_stream_listener(args)
 
         orchestrator = build_orchestrator(args)
-        outdir = _prepare_output_dir(args.output_dir)
+        outdir = prepare_output_dir(args.output_dir)
 
         run_existing_plan = args.execute and not args.prompt and not args.prompt_file
 
@@ -737,16 +284,16 @@ async def main(argv: Optional[list[str]] = None) -> int:
                 logging.error("--dry-run cannot be combined with executing a saved plan.")
                 return 2
 
-            selection, cancelled = _select_saved_plan(outdir)
+            selection, cancelled = select_saved_plan(outdir)
             if not selection:
                 return 0 if cancelled else 1
 
             plan = selection.plan
-            prompt = _ensure_prompt_for_plan(selection)
+            prompt = ensure_prompt_for_plan(selection)
             if not prompt:
                 return 1
 
-            plan_dir = _initialise_plan_storage(
+            plan_dir = initialise_plan_storage(
                 outdir,
                 plan,
                 prompt,
@@ -757,30 +304,25 @@ async def main(argv: Optional[list[str]] = None) -> int:
             should_run = True
         else:
             prompt = _load_prompt(args)
+            plan = orchestrator.plan(prompt)
 
             if args.dry_run:
-                plan = orchestrator.plan(prompt)
                 if args.show_plan:
-                    print(_plan_to_yaml_text(plan))
-                plan_dir = _initialise_plan_storage(outdir, plan, prompt)
-                _persist_plan_result(plan_dir, plan, {}, "json", status="dry-run")
+                    print(plan_to_yaml_text(plan))
+                plan_dir = initialise_plan_storage(outdir, plan, prompt)
+                persist_plan_result(plan_dir, plan, {}, "json", status="dry-run")
                 result_persisted = True
                 return 0
 
-            plan = orchestrator.plan(prompt)
-            plan_dir = _initialise_plan_storage(outdir, plan, prompt)
+            plan_dir = initialise_plan_storage(outdir, plan, prompt)
             built = orchestrator.build_crew(plan)
-
-            if args.execute:
-                should_run = True
-            else:
-                should_run = _prompt_for_execution(plan)
+            should_run = args.execute or _prompt_for_execution(plan)
 
         assert plan is not None and built is not None and prompt is not None and plan_dir is not None
 
         if args.show_plan:
             print("\n=== Plan (YAML) ===\n")
-            print(_plan_to_yaml_text(plan))
+            print(plan_to_yaml_text(plan))
 
         executed = False
         final_result = None
@@ -791,15 +333,13 @@ async def main(argv: Optional[list[str]] = None) -> int:
             final_result = run_out.get("result")
             executed = True
             saved_result = final_result
-            assert plan_dir is not None
-            _persist_plan_result(plan_dir, plan, saved_result, args.format, status="executed")
+            persist_plan_result(plan_dir, plan, saved_result, args.format, status="executed")
             result_persisted = True
         else:
             skip_message = "Execution skipped by user request."
             logging.info(skip_message)
             saved_result = {"status": "skipped", "message": skip_message}
-            assert plan_dir is not None
-            _persist_plan_result(plan_dir, plan, saved_result, args.format, status="skipped")
+            persist_plan_result(plan_dir, plan, saved_result, args.format, status="skipped")
             result_persisted = True
 
         if executed:
@@ -834,13 +374,13 @@ async def main(argv: Optional[list[str]] = None) -> int:
 
         return 0
 
-    except PlanGenerationError as e:
-        logging.error("%s", e)
+    except PlanGenerationError as exc:
+        logging.error("%s", exc)
         return 2
     except KeyboardInterrupt:
         if plan_dir:
             now_iso = datetime.now().isoformat(timespec="seconds")
-            _update_plan_metadata(
+            update_plan_metadata(
                 plan_dir,
                 status="cancelled",
                 executed=False,
@@ -849,21 +389,21 @@ async def main(argv: Optional[list[str]] = None) -> int:
             )
         logging.warning("Interrupted by user.")
         return 130
-    except Exception as e:
+    except Exception as exc:  # pragma: no cover - defensive guard
         if plan is not None:
-            _dump_plan_on_failure(plan, plan_dir)
+            dump_plan_on_failure(plan, plan_dir)
 
         if plan_dir and plan is not None and not result_persisted:
             now_iso = datetime.now().isoformat(timespec="seconds")
-            _update_plan_metadata(
+            update_plan_metadata(
                 plan_dir,
                 status="error",
                 executed=False,
-                error=str(e),
+                error=str(exc),
                 updated_at=now_iso,
                 completed_at=now_iso,
             )
-        logging.exception("Unhandled error: %s", e)
+        logging.exception("Unhandled error: %s", exc)
         return 1
